Move getHealthVariant to package
