+++ v2/lib/getHealthVariant/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@snx-v2/getHealthVariant",
+  "private": true,
+  "main": "build/index.js",
+  "version": "0.0.0",
+  "scripts": {
+    "build": "tsc"
+  },
+  "devDependencies": {
+    "typescript": "^4.8.2"
+  }
+}
