+++ v2/lib/getHealthVariant/src/getHealthVariant.ts
