+++ v2/lib/getHealthVariant/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig.base.module.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "outDir": "./build",
+    "rootDir": "./src"
+  },
+  "include": ["src"]
+}
