+++ v2/ui/package.json
@@ -17,7 +17,9 @@
     "test": "NODE_ENV=test jest",
     "test:unit": "yarn test",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "standalone-install": "yarn workspaces focus '@synthetixio/v2-ui'",
+    "standalone-build": "yarn workspaces foreach --topological-dev --recursive --verbose --from '@synthetixio/v2-ui' run build"
   },
   "dependencies": {
     "@artsy/fresnel": "^1.8.0",
